WIP: Add failing tests for bad log lines
I've copied the bad log line (and the surrounding two) from
the server on ocelot, and made it into a fixture. Then written
a test for the exim log loading method to make sure it doesn't
error. The thing is, it seems to work fine locally :(. I'm not
sure if it's a ruby-version thing, something I've done wrong
getting the line out of the logfile, or something else.